Refactor chatroom recipe for docs
